Adds notes for build and run

+++ ansible/clean.sh
@@ -1,22 +1,32 @@
 #!/bin/bash
 #
-# Removes all containers and images associated with the app 
+# Removes containers and images associated with the app.
+#
+# By default will remove all containers and images.
+# Pass 'container' to only remove containers. 
+# Pass 'images' to remove images only.
 #
 
-echo "Removing all Galaxy containers and images..."
-
-containers=$(docker ps -a --format "{{.Names}}" | grep -e django -e elastic -e postgres -e rabbit -e memcache -e gulp | wc -l | tr -d '[[:space:]]')
-if [ ${containers} -gt 0 ]; then 
-    docker rm --force $(docker ps -a --format "{{.Names}}" | grep -e django -e elastic -e postgres -e rabbit -e memcache -e gulp)
-else
-    echo "No Galaxy containers found"
+if [ $# -eq 0 ] || [ "$1" = "containers" ]; then
+    echo "Removing Galaxy containers..."
+    containers=$(docker ps -a --format "{{.Names}}" | grep -e django -e elastic -e postgres -e rabbit -e memcache -e gulp | wc -l | tr -d '[[:space:]]')
+    if [ ${containers} -gt 0 ]; then 
+        docker rm --force $(docker ps -a --format "{{.Names}}" | grep -e django -e elastic -e postgres -e rabbit -e memcache -e gulp)
+    else
+        echo "No Galaxy containers found"
+    fi
 fi
 
-images=$(docker images -a --format "{{.Repository}}:{{.Tag}}" | grep galaxy | wc -l | tr -d '[[:space:]]')
-if [ ${images} -gt 0 ]; then
-    docker rmi --force $(docker images -a --format "{{.Repository}}:{{.Tag}}" | grep galaxy)
-else
-    echo "No Galaxy images found"
+if [ $# -eq 0 ] || [ "$1" = "images" ]; then
+    echo "Removing Galaxy images..."
+    images=$(docker images -a --format "{{.Repository}}:{{.Tag}}" | grep galaxy | wc -l | tr -d '[[:space:]]')
+    if [ ${images} -gt 0 ]; then
+        docker rmi --force $(docker images -a --format "{{.Repository}}:{{.Tag}}" | grep galaxy)
+    else
+        echo "No Galaxy images found"
+    fi
 fi
 
-rm -rf ~/.galaxy
+if [ $# -eq 0 ]; then
+    rm -rf ~/.galaxy
+fi
